backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Cache & Queue
redis==5.2.1
celery[redis]==5.4.0

# Vector Store
qdrant-client==1.12.1

# Agent Framework
langgraph==0.2.60
langchain==0.3.14
langchain-openai==0.3.0
langchain-core==0.3.29

# Federal Corpus Parsing
lxml==5.3.0
httpx==0.28.1

# Document Parsing
PyMuPDF==1.25.3
python-docx==1.1.2
pdfplumber==0.11.4

# Observability
prometheus-client==0.21.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
